// nutrition-assistant-front/src/app/model/food/nutrients/amino-acid.enum.ts
export enum AminoAcid {

    CYSTEINE = 'Cysteine',
    HISTIDINE = 'Histidine',
    ISOLEUCINE = 'Isoleucine',
    LEUCINE = 'Leucine',
    LYSINE = 'Lysine',
    METHIONINE = 'Methionine',
    PHENYLALANINE = 'Phenylalanine',
    THREONINE = 'Threonine',
    TRYPTOPHAN = 'Tryptophan',
    TYROSINE = 'Tyrosine',
    VALINE = 'Valine'

}