// nutrition-assistant-front/src/app/model/food/nutrients/vitamin.enum.ts
export enum Vitamin {

    A = 'A',
    B1 = 'B1 (Thiamine)',
    B2 = 'B2 (Riboflavin)',
    B3 = 'B3 (Niacin)',
    B4 = 'B4 (Choline)',
    B5 = 'B5 (Panthotenic acid)',
    B6 = 'B6 (Pyridoxine)',
    B7 = 'B7 (Biotin)',
    B12 = 'B12 (Cobalamin)',
    C = 'C',
    D = 'D',
    E = 'E',
    FOLATE = 'Folate',
    K = 'K'

}