// nutrition-assistant-front/src/app/model/food/nutrients/mineral.enum.ts
export enum Mineral {

    CALCIUM = 'Calcium',
    COPPER = 'Copper',
    FLUORIDE = 'Fluoride',
    IRON = 'Iron',
    MAGNESIUM = 'Magnesium',
    MANGANESE = 'Manganese',
    PHOSPHORUS = 'Phosphorus',
    POTASSIUM = 'Potassium',
    SELENIUM = 'Selenium',
    SODIUM = 'Sodium',
    ZINC = 'Zinc'

}